+++ file/file_prefetch_buffer.cc
@@ -107,6 +107,7 @@ Status FilePrefetchBuffer::Read(const IOOptions& opts,
 }
 
 Status FilePrefetchBuffer::ReadAsync(const IOOptions& opts,
+                                     Env::IOPriority rate_limiter_priority,
                                      RandomAccessFileReader* reader,
                                      uint64_t read_len, uint64_t chunk_len,
                                      uint64_t rounddown_start, uint32_t index) {
@@ -119,7 +120,7 @@ Status FilePrefetchBuffer::ReadAsync(const IOOptions& opts,
   req.offset = rounddown_start + chunk_len;
   req.result = result;
   req.scratch = bufs_[index].buffer_.BufferStart() + chunk_len;
-  Status s = reader->ReadAsync(req, opts, fp,
+  Status s = reader->ReadAsync(req, opts, rate_limiter_priority, fp,
                                /*cb_arg=*/nullptr, &io_handle_, &del_fn_,
                                /*aligned_buf=*/nullptr);
   req.status.PermitUncheckedError();
@@ -373,7 +374,8 @@ Status FilePrefetchBuffer::PrefetchAsyncInternal(
     bufs_[second].offset_ = rounddown_start2;
     assert(roundup_len2 >= chunk_len2);
     uint64_t read_len2 = static_cast<size_t>(roundup_len2 - chunk_len2);
-    ReadAsync(opts, reader, read_len2, chunk_len2, rounddown_start2, second)
+    ReadAsync(opts, rate_limiter_priority, reader, read_len2, chunk_len2,
+              rounddown_start2, second)
         .PermitUncheckedError();
   }
 
@@ -558,6 +560,7 @@ void FilePrefetchBuffer::PrefetchAsyncCallback(const FSReadRequest& req,
 }
 
 Status FilePrefetchBuffer::PrefetchAsync(const IOOptions& opts,
+                                         Env::IOPriority rate_limiter_priority,
                                          RandomAccessFileReader* reader,
                                          uint64_t offset, size_t n,
                                          Slice* result) {
@@ -629,7 +632,8 @@ Status FilePrefetchBuffer::PrefetchAsync(const IOOptions& opts,
 
   size_t read_len = static_cast<size_t>(roundup_len - chunk_len);
 
-  s = ReadAsync(opts, reader, read_len, chunk_len, rounddown_start, second);
+  s = ReadAsync(opts, rate_limiter_priority, reader, read_len, chunk_len,
+                rounddown_start, second);
 
   if (!s.ok()) {
     return s;