+++ file/file_prefetch_buffer.h
@@ -175,8 +175,10 @@ class FilePrefetchBuffer {
   //
   // If data already exist in the buffer, it will return Status::OK, otherwise
   // it will send asynchronous request and return Status::TryAgain.
-  Status PrefetchAsync(const IOOptions& opts, RandomAccessFileReader* reader,
-                       uint64_t offset, size_t n, Slice* result);
+  Status PrefetchAsync(const IOOptions& opts,
+                       Env::IOPriority rate_limiter_priority,
+                       RandomAccessFileReader* reader, uint64_t offset,
+                       size_t n, Slice* result);
 
   // Tries returning the data for a file read from this buffer if that data is
   // in the buffer.
@@ -274,9 +276,10 @@ class FilePrefetchBuffer {
               Env::IOPriority rate_limiter_priority, uint64_t read_len,
               uint64_t chunk_len, uint64_t rounddown_start, uint32_t index);
 
-  Status ReadAsync(const IOOptions& opts, RandomAccessFileReader* reader,
-                   uint64_t read_len, uint64_t chunk_len,
-                   uint64_t rounddown_start, uint32_t index);
+  Status ReadAsync(const IOOptions& opts, Env::IOPriority rate_limiter_priority,
+                   RandomAccessFileReader* reader, uint64_t read_len,
+                   uint64_t chunk_len, uint64_t rounddown_start,
+                   uint32_t index);
 
   // Copy the data from src to third buffer.
   void CopyDataToBuffer(uint32_t src, uint64_t& offset, size_t& length);